+++ client/rpc/Cargo.toml
@@ -46,6 +46,7 @@ assert_matches = "1.3.0"
 lazy_static = "1.4.0"
 sc-block-builder = { version = "0.10.0-dev", path = "../block-builder" }
 sc-network = { version = "0.10.0-dev", path = "../network" }
+sc-network-common = { version = "0.10.0-dev", path = "../network/common" }
 sc-transaction-pool = { version = "4.0.0-dev", path = "../transaction-pool" }
 sp-consensus = { version = "0.10.0-dev", path = "../../primitives/consensus/common" }
 tokio = "1.17.0"
